refine autoround format
