+++ auto_round/inference/auto_quantizer.py
@@ -1,0 +1,417 @@
+# Copyright (c) 2024 Intel Corporation
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+# coding=utf-8
+# Copyright 2023 HuggingFace Inc. team and GPTQ and AutoGPTQ authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import gc
+import importlib.util
+import warnings
+from dataclasses import dataclass
+from logging import getLogger
+from typing import Any, Dict, Optional, Tuple, Union
+
+import torch
+import torch.nn as nn
+from packaging import version
+from transformers.modeling_utils import PreTrainedModel
+from transformers.pytorch_utils import Conv1D
+from transformers.quantizers import AutoQuantizationConfig, HfQuantizer
+from transformers.quantizers.auto import AUTO_QUANTIZER_MAPPING
+from transformers.utils.quantization_config import AwqConfig, GPTQConfig, QuantizationConfigMixin, QuantizationMethod
+from auto_round.utils import (get_module, set_module, is_hpu_supported, get_block_names,
+                              find_matching_blocks)
+from auto_round.inference.convert_model import convert_hf_model
+
+from auto_round.inference.backend import get_layer_backend, dynamic_import_inference_linear
+
+from auto_round.inference.backend import BackendInfos
+from transformers.utils.versions import require_version
+from enum import Enum
+from tqdm import tqdm
+
+logger = getLogger(__name__)
+import sys
+
+if sys.version_info < (3, 8):
+    import importlib_metadata
+else:
+    import importlib.metadata as importlib_metadata
+
+AUTOROUND_MINIMUM_VERSION = version.parse("0.2")
+
+
+def _is_package_available(pkg_name: str, return_version: bool = False) -> Union[Tuple[bool, str], bool]:
+    # Check we're not importing a "pkg_name" directory somewhere but the actual library by trying to grab the version
+    try:  ##TODO remove it later
+        import auto_round
+        return True, auto_round.__version__
+    except:
+        pass
+
+    package_exists = importlib.util.find_spec(pkg_name) is not None
+    package_version = "N/A"
+    if package_exists:
+        try:
+            package_version = importlib.metadata.version(pkg_name)
+            package_exists = True
+        except importlib.metadata.PackageNotFoundError:
+            package_exists = False
+    if return_version:
+        return package_exists, package_version
+    else:
+        return package_exists
+
+
+_auto_round_available = _is_package_available("auto_round")
+
+
+def is_auto_round_available():
+    try:
+        import auto_round
+        return True
+    except:
+        pass
+    if _auto_round_available:
+        version_autoround = version.parse(importlib_metadata.version("auto_round"))
+        if AUTOROUND_MINIMUM_VERSION < version_autoround:
+            return True
+        else:
+            raise ImportError(
+                f"Found an incompatible version of auto-round. Found version {version_autoround},"
+                f" but only version above {AUTOROUND_MINIMUM_VERSION} are supported"
+            )
+
+
+
+class AutoHfQuantizer:
+    """The Auto-HF quantizer class that takes care of automatically instantiating to the correct
+    `HfQuantizer` given the `QuantizationConfig`."""
+
+    @classmethod
+    def from_config(cls, quantization_config: Union[QuantizationConfigMixin, Dict], **kwargs):
+        # Convert it to a QuantizationConfig if the q_config is a dict
+        if isinstance(quantization_config, dict):
+            if "auto-round" in quantization_config["quant_method"]:
+                quantization_config = AutoRoundConfig.from_dict(quantization_config)
+            else:
+                quantization_config = AutoQuantizationConfig.from_dict(quantization_config)  # pylint: disable=E1101
+        quant_method = quantization_config.quant_method
+
+        # Again, we need a special care for bnb as we have a single quantization config
+        # class for both 4-bit and 8-bit quantization
+        if quant_method == QuantizationMethod.BITS_AND_BYTES:
+            if quantization_config.load_in_8bit:
+                quant_method += "_8bit"
+            else:
+                quant_method += "_4bit"
+
+        if quant_method not in AUTO_QUANTIZER_MAPPING.keys() and "auto-round" not in quant_method:
+            raise ValueError(
+                f"Unknown quantization type, got {quant_method} - supported types are:"
+                f" {list(AUTO_QUANTIZER_MAPPING.keys())}"
+            )
+        if "auto-round" in quant_method or is_hpu_supported():  # pragma: no cover
+            target_cls = AutoRoundQuantizer
+        else:
+            target_cls = AUTO_QUANTIZER_MAPPING[quant_method]
+
+        return target_cls(quantization_config, **kwargs)
+
+    @classmethod
+    def from_pretrained(cls, pretrained_model_name_or_path, **kwargs):
+        quantization_config = AutoQuantizationConfig.from_pretrained(pretrained_model_name_or_path, **kwargs)
+        return cls.from_config(quantization_config)
+
+    @classmethod
+    def merge_quantization_configs(
+            cls,
+            quantization_config: Union[dict, QuantizationConfigMixin],
+            quantization_config_from_args: Optional[QuantizationConfigMixin],
+    ):
+        """Handles situations where both quantization_config
+        from args and quantization_config from model config are present."""
+        if quantization_config_from_args is not None:
+            warning_msg = (
+                "You passed `quantization_config` or equivalent parameters to "
+                "`from_pretrained` but the model you're loading"
+                " already has a `quantization_config` attribute. The `quantization_config` from the model will be used."
+            )
+        else:
+            warning_msg = ""
+
+        loading_attr_dict = quantization_config_from_args.get_loading_attributes() \
+            if quantization_config_from_args is not None else None
+        if isinstance(quantization_config, dict):
+            if "auto-round" in quantization_config["quant_method"]:
+                quantization_config = AutoRoundConfig.from_dict(quantization_config)
+            else:
+                if isinstance(quantization_config_from_args, (AutoRoundConfig)):
+                    logger.info(f"Loading quantized model in auto_round format.")
+                    tmp_backend = quantization_config["quant_method"]
+                    if "auto-round" not in tmp_backend and "gptq" not in tmp_backend and "awq" not in tmp_backend:
+                        logger.error("could not convert to auto_round format, currently only supports `gptq`,`awq` or "
+                                     "`auto-round` format")
+                        exit(-1)
+                    target_backend = quantization_config["backend"] if "backend" in quantization_config else "auto"
+                    if loading_attr_dict is not None and "backend" in loading_attr_dict:
+                        target_backend = loading_attr_dict["backend"]
+                        loading_attr_dict.pop("backend")
+                    if "auto_round" not in target_backend:
+                        target_backend = f"auto_round:{tmp_backend}"  #
+                    quantization_config = AutoRoundConfig.from_dict(quantization_config)
+                    setattr(quantization_config, "backend", target_backend)
+                else:
+                    quantization_config = AutoQuantizationConfig.from_dict(quantization_config)  # pylint: disable=E1101
+
+        if isinstance(quantization_config,
+                      (GPTQConfig, AwqConfig, AutoRoundConfig)) and quantization_config_from_args is not None:
+            # special case for GPTQ / AWQ config collision
+
+            for attr, val in loading_attr_dict.items():
+                setattr(quantization_config, attr, val)
+            warning_msg += (
+                f"However, loading attributes (e.g. {list(loading_attr_dict.keys())}) "
+                f"will be overwritten with the one you passed to `from_pretrained`. The rest will be ignored."
+            )
+
+        if warning_msg != "":
+            warnings.warn(warning_msg)
+
+        return quantization_config
+
+    @staticmethod
+    def supports_quant_method(quantization_config_dict):
+        from transformers.quantizers.auto import AUTO_QUANTIZATION_CONFIG_MAPPING
+        AUTO_QUANTIZATION_CONFIG_MAPPING['intel/auto-round'] = AutoRoundConfig
+        AUTO_QUANTIZATION_CONFIG_MAPPING['intel/auto_round'] = AutoRoundConfig
+        quant_method = quantization_config_dict.get("quant_method", None)
+        if quantization_config_dict.get("load_in_8bit", False) or quantization_config_dict.get("load_in_4bit", False):
+            suffix = "_4bit" if quantization_config_dict.get("load_in_4bit", False) else "_8bit"
+            quant_method = QuantizationMethod.BITS_AND_BYTES + suffix
+        elif quant_method is None:
+            raise ValueError(
+                "The model's quantization config from the arguments has no `quant_method` attribute." \
+                "Make sure that the model has been correctly quantized"
+            )
+
+        if quant_method not in AUTO_QUANTIZATION_CONFIG_MAPPING.keys():
+            logger.warning(
+                f"Unknown quantization type, got {quant_method} - supported types are:"
+                f" {list(AUTO_QUANTIZER_MAPPING.keys())}. Hence, we will skip the quantization. "
+                "To remove the warning, you can delete the quantization_config attribute in config.json"
+            )
+            return False
+        return True
+
+
+class AutoRoundQuantizationMethod(str, Enum):
+    AutoRound = "intel/auto-round"
+
+
+@dataclass
+class AutoRoundConfig(QuantizationConfigMixin):
+    """This is a wrapper class about all possible attributes and features that you can play with a model that has been
+    loaded AutoRound quantization.
+
+    Args:
+        bits (`int`):
+            The number of bits to quantize to, supported numbers are (2, 3, 4, 8).
+        tokenizer (`str` or `PreTrainedTokenizerBase`, *optional*):
+            The tokenizer used to process the dataset. You can pass either:
+                - A custom tokenizer object.
+                - A string, the *model id* of a predefined tokenizer hosted inside a model repo on huggingface.co.
+                - A path to a *directory* containing vocabulary files required by the tokenizer, for instance saved
+                    using the [`~PreTrainedTokenizer.save_pretrained`] method, e.g., `./my_model_directory/`.
+    """
+
+    def __init__(
+            self,
+            bits: int = 4,
+            tokenizer: Any = None,
+            dataset: str = None,
+            group_size: int = 128,
+            sym: bool = False,
+            backend="auto",
+            layer_config: dict = None,
+            **kwargs,
+    ):
+
+        self.bits = bits
+        self.tokenizer = tokenizer
+        self.dataset = dataset
+        self.group_size = group_size
+        self.sym = sym
+        self.target_backend = backend
+        self.layer_config = layer_config
+        if kwargs is not None:
+            for key in kwargs.keys():
+                setattr(self, key, kwargs[key])
+        self.quant_method = AutoRoundQuantizationMethod.AutoRound
+        self.post_init()
+
+    def post_init(self):
+        r"""Safety checker that arguments are correct."""
+        if self.bits not in [2, 4, 8]:
+            raise ValueError(f"Only support quantization to [2,4,8] bits but found {self.bits}")
+        if self.group_size != -1 and self.group_size <= 0:
+            raise ValueError("group_size must be greater than 0 or equal to -1")
+
+    def get_loading_attributes(self):
+        loading_attibutes_dict = {"target_backend": self.target_backend}
+        return loading_attibutes_dict
+
+    def to_dict(self):
+        config_dict = super().to_dict()
+        return config_dict
+
+
+class AutoRoundQuantizer(HfQuantizer):
+    """Quantizer of the AutoRound method, currently only triton and exllamav2 backend has been supported."""
+
+    requires_calibration = False
+    required_packages = ["auto_round"]
+    optimum_quantizer = None
+
+    def __init__(self, quantization_config: QuantizationConfigMixin, **kwargs):
+        super().__init__(quantization_config, **kwargs)
+
+    def validate_environment(self, *args, **kwargs):
+        if not is_auto_round_available():
+            raise ImportError("Loading a AutoRound quantized model requires auto-round library (`pip install "
+                              "auto-round`)")
+        else:
+            try:
+                import auto_round
+                autoround_version = version.parse(auto_round.__version__)
+            except:
+                autoround_version = version.parse(importlib.metadata.version("auto_round"))
+            if autoround_version < version.parse("0.2.0"):
+                raise ImportError("You need a version of auto_round > 0.2.0 to use AutoRound: `pip install --upgrade "
+                                  "auto-round` or install from source")
+
+    def update_torch_dtype(self, torch_dtype: "torch.dtype") -> "torch.dtype":
+        if torch_dtype is None:
+            torch_dtype = torch.float16
+        return torch_dtype
+
+
+
+
+
+    # def cpu_post_init(self, model):
+    #     message = "Repacking to CPU format"
+    #     from auto_round_extension.qbits import qbits_qlinear_classes, qbits_awq_classes
+    #     from auto_round_extension.ipex import ipex_qlinear_classes
+    #     cpu_layers = tuple(list(qbits_qlinear_classes) + list(ipex_qlinear_classes) + list(qbits_awq_classes))
+    #     layers = []  ## ipex post_init  will add one more layer
+    #     for n, m in model.named_modules():
+    #         if isinstance(m, cpu_layers):
+    #             layers.append((n, m))
+    #     for n, layer in tqdm(layers, desc=message, total=len(layers),
+    #                          leave=True):
+    #         layer.post_init()
+    #     return model
+    #
+    # def xpu_post_init(self, model):
+    #     message = "Repacking to XPU format"
+    #     from auto_round_extension.ipex import ipex_qlinear_classes
+    #     cpu_layers = tuple(list(ipex_qlinear_classes))
+    #     layers = []  ## ipex post_init  will add one more layer
+    #     for n, m in model.named_modules():
+    #         if isinstance(m, cpu_layers):
+    #             layers.append((n, m))
+    #     for n, layer in tqdm(layers, desc=message, total=len(layers),
+    #                          leave=True):
+    #         layer.post_init()
+    #     return model
+
+
+    def post_init_model(self, model):
+        """Post-initialization that require device information, for example buffers initialization on device.
+
+        Args:
+            model (`nn.Module`):
+                The input model
+        """
+
+        class StoreAttr(object):
+            pass
+
+        model.quantize_config = StoreAttr()
+        if hasattr(self,"used_backend_info") and self.used_backend_info["used_autogptq"]:
+            from auto_gptq.modeling._utils import autogptq_post_init as gptq_post_init
+            model = gptq_post_init(model, use_act_order=False)
+        if hasattr(self, "used_backend_info") and self.used_backend_info["used_ipex"] or self.used_backend_info["used_qbits"]:
+            logger.info("repacking to CPU/XPU format") ## TODO use tqdm to show progress
+            for n,m in model.named_modules():
+                if hasattr(m, "post_init") and callable(getattr(m, "post_init")):
+                    m.post_init()
+        # if self.need_marlin_repacking:
+        #     require_version("gptqmodel",
+        #                     "marlin format requires gptqmodel to be installed, "
+        #                     "`pip install -v gptqmodel --no-build-isolation `")
+        #     self.repack_marlin(model)
+        # from auto_round_extension.cuda.post_init import autoround_post_init
+        # model = autoround_post_init(model)
+        # # there are no side-effects after call qbits_post_init when model quant-type not equal to qbits.
+        # if self.target_device == "cpu":
+        #     model = self.cpu_post_init(model)
+        # elif self.target_device == "xpu":
+        #     model = self.xpu_post_init(model)
+        #
+        # return model
+
+    def _process_model_before_weight_loading(self, model: "PreTrainedModel", **kwargs):
+        if model.__class__.main_input_name != "input_ids":
+            logger.warning("We can only quantize pure text models and " \
+                           "certain types(Llava/Qwen-VL/Phi-3-vision) of multimodal models.")
+
+        if self.pre_quantized:
+            model,used_backend_info = convert_hf_model(model)
+            self.used_backend_info = used_backend_info
+
+    def _process_model_after_weight_loading(self, model: "PreTrainedModel", **kwargs):
+        if self.pre_quantized:
+            model = self.post_init_model(model)
+        else:
+            raise NotImplementedError
+
+    @property
+    def is_trainable(self, model: Optional["PreTrainedModel"] = None):
+        return True
+
+    @property
+    def is_serializable(self):
+        return True
+
+
+import transformers
+
+if version.parse(transformers.__version__) < version.parse("4.38.0"):
+    logger.error("Please upgrade transformers>=4.38.0 to support lm-head quantization")
+
+transformers.quantizers.auto.AutoHfQuantizer = AutoHfQuantizer
+transformers.modeling_utils.AutoHfQuantizer = AutoHfQuantizer
